+++ Class/Toast.cs
@@ -16,7 +16,7 @@ namespace Mr_Squirrely_Converters.Class {
             });
         }
         #region Messages
-        internal static void NoUpdate() => _Notifier.ShowSuccess("There is no update!");// Shows if there is no update
+        internal static void NoUpdate() => _Notifier.ShowSuccess("There is no update!"); // Shows if there is no update
         internal static void Update() => _Notifier.ShowInformation($"There is an update. Your version: {Utils._CurrentVersion} Updated version: {Utils._UpdateVerstion}"); //This message shows when there is an update
         internal static void UpdateCheckFail() => _Notifier.ShowWarning("Failed to check for update. Please try again."); //In case checking for the update fails
         internal static void BetaRelease() => _Notifier.ShowInformation("This is a beta release so some things are not finished."); //Beta release notice
@@ -24,7 +24,7 @@ namespace Mr_Squirrely_Converters.Class {
         internal static void AlreadyConverting() => _Notifier.ShowWarning("Already Converting"); //Already converting message
         internal static void SettingsSaved() => _Notifier.ShowSuccess("Settings were saved!"); //Settings saved message
         internal static void SettingsReset() => _Notifier.ShowInformation("Settings reset, make sure you save them!"); //Reset message and a reminder to save them
-        internal static void VideoMessage() => _Notifier.ShowWarning("Video conversion can take a long time to finish. Currently I do not show progress. It is not recommended that you convert multiple videos at once.");//Message for video
+        internal static void VideoMessage() => _Notifier.ShowWarning("Video conversion can take a long time to finish. Currently I do not show progress. It is not recommended that you convert multiple videos at once."); //Message for video
         internal static void VideoMessage2() => _Notifier.ShowWarning("Currently converting of videos is disabled.");
         #endregion
