+++ Views/MainWindow.xaml.cs
@@ -8,7 +8,6 @@ namespace Mr_Squirrely_Converters {
     /// Interaction logic for MainWindow.xaml
     /// </summary>
     public partial class MainWindow {
-
         public MainWindow() {
             InitializeComponent();
             Toast.CreateNotifier();
@@ -28,6 +27,5 @@ namespace Mr_Squirrely_Converters {
         private void SettingsMenu_Click(object sender, RoutedEventArgs e) => Utils._MainWindow.Content = Utils._SettingsPage;
         private void ClearMenu_Click(object sender, RoutedEventArgs e) => Utils.Clear();
         private void UpdateMenu_Click(object sender, RoutedEventArgs e) => Utils.CheckForUpdate(true);
-
     }
 }
